refactor(dhcp): extract DHCPv4 client to dhcpv4.go
Continue the #6430 split by moving the DHCPv4 client out of dhcp.go
into a new sibling dhcpv4.go: the runDHCPv4() DORA / T1-renew / T2-rebind
run loop, the errDHCPNAK sentinel and abandonLeaseAfterNAK() revocation
handler (RFC 2131 §4.4.5), doDHCPv4() single-exchange primitive,
leaseFromACKv4() ACK parsing (including the degenerate-subnet-mask
refuse), and classlessStaticRoutes() RFC 3442 option-121/249 parsing.

Pure code-motion: whole functions and the errDHCPNAK var move verbatim
with unchanged signatures and bodies; only the per-file package clause
and (goimports-pruned) import block differ. The v6 sentinel
errV6AddrInvalidated, which happened to sit inside the v4 region of the
original file, stays with the DHCPv6 client (moved in the next commit).

Validation: go build ./pkg/dhcp/, go vet ./pkg/dhcp/ clean.
